+++ src/components/planner/Planner.tsx
@@ -10,6 +10,8 @@
  * Be very careful when adding @dnd-kit presets (eg. @dnd-kit/sortable) or new data sources
  * Adding DndContext hooks (eg. useDraggable) without considering 'data' property will cause unexpected behaviors
  */
+import { trpc } from '@/utils/trpc';
+import { useTaskQueue } from '@/utils/useTaskQueue';
 import {
   DndContext,
   DragOverlay,
@@ -19,6 +21,7 @@ import {
   useSensors,
 } from '@dnd-kit/core';
 import React, { useState } from 'react';
+import { toast } from 'react-toastify';
 
 import CourseSelectorContainer from './Sidebar/Sidebar';
 import { SidebarCourseItem } from './Sidebar/SidebarCourseItem';
@@ -31,42 +34,23 @@ import {
   DragEventOriginData,
   DraggableCourse,
   Semester,
-  ToastMessage,
 } from './types';
+import { createNewYear } from '@/utils/utilFunctions';
 
 /** PlannerTool Props */
 export interface PlannerProps {
   degreeRequirements: DegreeRequirementGroup[];
   semesters: Semester[];
-  /** Called when course moved from course list -> semester */
-  onAddCourseToSemester?: (
-    targetSemester: Semester,
-    newCourse: DraggableCourse,
-  ) => Promise<ToastMessage>;
-  /** Called when course removed from semester */
-  onRemoveCourseFromSemester: (
-    targetSemester: Semester,
-    courseToRemove: DraggableCourse,
-  ) => Promise<ToastMessage>;
-  /** Called when courese moved from semester -> semester */
-  onMoveCourseFromSemesterToSemester?: (
-    originSemester: Semester,
-    destinationSemester: Semester,
-    courseToMove: DraggableCourse,
-  ) => Promise<ToastMessage>;
-  onRemoveYear: () => Promise<void>;
-  onAddYear: () => Promise<void>;
+  planId: string;
+  setSemesters: React.Dispatch<React.SetStateAction<Semester[]>>;
 }
 
 /** Controlled wrapper around course list and semester tiles */
 export default function Planner({
   degreeRequirements,
   semesters,
-  onAddCourseToSemester,
-  onRemoveCourseFromSemester,
-  onMoveCourseFromSemesterToSemester,
-  onRemoveYear,
-  onAddYear,
+  setSemesters,
+  planId,
 }: PlannerProps): JSX.Element {
   // Course that is currently being dragged
   const [activeCourse, setActiveCourse] = useState<ActiveDragData | null>(null);
@@ -80,6 +64,243 @@ export default function Planner({
       },
     }),
   );
+
+  // Extend Course object to contain fields for prereq validation & other stuff ig
+
+  const utils = trpc.useContext();
+
+  const { addTask } = useTaskQueue({ shouldProcess: true });
+
+  const addCourse = trpc.plan.addCourseToSemester.useMutation({
+    async onSuccess() {
+      await utils.plan.getPlanById.invalidate(planId);
+    },
+  });
+
+  const removeCourse = trpc.plan.removeCourseFromSemester.useMutation({
+    async onSuccess() {
+      await utils.plan.getPlanById.invalidate(planId);
+    },
+  });
+
+  const moveCourse = trpc.plan.moveCourseFromSemester.useMutation({
+    async onSuccess() {
+      await utils.plan.getPlanById.invalidate(planId);
+    },
+  });
+
+  const createYear = trpc.plan.addYear.useMutation({
+    async onSuccess() {
+      await utils.plan.getPlanById.invalidate(planId);
+    },
+  });
+
+  const deleteYear = trpc.plan.deleteYear.useMutation({
+    async onSuccess() {
+      await utils.plan.getPlanById.invalidate(planId);
+    },
+  });
+
+  const handleYearCreate = async ({ semesterIds }: { [key: string]: string[] }) => {
+    try {
+      await toast.promise(
+        createYear.mutateAsync({
+          planId,
+          semesterIds: semesterIds.map((id) => id),
+        }),
+        {
+          pending: 'Creating year...',
+          success: 'Year created!',
+          error: 'Error creating year',
+        },
+        {
+          autoClose: 1000,
+        },
+      );
+    } catch (error) {
+      console.error(error);
+    }
+  };
+  const handleYearDelete = async () => {
+    try {
+      // TODO: Handle deletion errors
+
+      await toast.promise(
+        deleteYear.mutateAsync(planId),
+        {
+          pending: 'Deleting year...',
+          success: 'Year deleted!',
+          error: 'Error deleting year',
+        },
+        {
+          autoClose: 1000,
+        },
+      );
+    } catch (error) {}
+  };
+
+  const handleAddCourse = async ({ semesterId, courseName }: { [key: string]: string }) => {
+    try {
+      await toast.promise(
+        addCourse.mutateAsync({ planId, semesterId, courseName }),
+        {
+          pending: 'Adding course ' + courseName + '...',
+          success: 'Added course ' + courseName + '!',
+          error: 'Error in adding ' + courseName,
+        },
+        {
+          autoClose: 1000,
+        },
+      );
+    } catch (error) {}
+  };
+
+  const handleRemoveCourse = async ({
+    semesterId: semesterId,
+    courseName: courseName,
+  }: {
+    [key: string]: string;
+  }) => {
+    try {
+      await toast.promise(
+        removeCourse.mutateAsync({ planId, semesterId, courseName }),
+        {
+          pending: 'Removing course ' + courseName + '...',
+          success: 'Removed course ' + courseName + '!',
+          error: 'Error in removing ' + courseName,
+        },
+        {
+          autoClose: 1000,
+        },
+      );
+    } catch (error) {}
+  };
+
+  const handleMoveCourse = async ({
+    oldSemesterId,
+    newSemesterId,
+    courseName,
+  }: {
+    [key: string]: string;
+  }) => {
+    try {
+      await toast.promise(
+        moveCourse.mutateAsync({ planId, oldSemesterId, newSemesterId, courseName }),
+        {
+          pending: 'Moving course ' + courseName + '...',
+          success: 'Moved course ' + courseName + '!',
+          error: 'Error while moving ' + courseName,
+        },
+        {
+          autoClose: 1000,
+        },
+      );
+    } catch (error) {
+      console.error(error);
+    }
+  };
+
+  const handleAddYear = () => {
+    const newYear: Semester[] = createNewYear(
+      semesters.length ? semesters[semesters.length - 1].code : { semester: 'u', year: 2022 },
+    );
+    const semesterIds = newYear.map((sem) => sem.id);
+    setSemesters([...semesters, ...newYear]);
+    addTask({
+      func: handleYearCreate,
+      args: { semesterIds: semesterIds.map((id) => id.toString()) },
+    });
+  };
+
+  const handleRemoveYear = () => {
+    setSemesters(semesters.filter((_, idx) => idx < semesters.length - 3));
+    addTask({ func: handleYearDelete, args: {} });
+  };
+
+  const handleRemoveCourseFromSemester = (
+    targetSemester: Semester,
+    targetCourse: DraggableCourse,
+  ) => {
+    setSemesters((semesters) =>
+      semesters.map((semester) => {
+        if (semester.id === targetSemester.id) {
+          return {
+            ...semester,
+            courses: semester.courses.filter((course) => course.id !== targetCourse.id),
+          };
+        }
+
+        return semester;
+      }),
+    );
+
+    const semesterId = targetSemester.id.toString();
+    const courseName = targetCourse.code;
+    addTask({ func: handleRemoveCourse, args: { semesterId, courseName } });
+  };
+
+  const handleAddCourseToSemester = (targetSemester: Semester, newCourse: DraggableCourse) => {
+    // check for duplicate course
+    const isDuplicate = Boolean(
+      targetSemester.courses.find((course) => course.code === newCourse.code),
+    );
+    if (isDuplicate) {
+      toast.warn(
+        `You're already taking ${newCourse.code} in ${targetSemester.code.year}${targetSemester.code.semester}`,
+      );
+      return;
+    }
+
+    setSemesters((semesters) =>
+      semesters.map((semester) =>
+        semester.id === targetSemester.id
+          ? { ...semester, courses: [...semester.courses, newCourse] }
+          : semester,
+      ),
+    );
+    const semesterId = targetSemester.id.toString();
+    const courseName = newCourse.code;
+    addTask({ func: handleAddCourse, args: { semesterId, courseName } });
+  };
+
+  const handleMoveCourseFromSemesterToSemester = (
+    originSemester: Semester,
+    destinationSemester: Semester,
+    courseToMove: DraggableCourse,
+  ) => {
+    // check for duplicate course
+    const isDuplicate = Boolean(
+      destinationSemester.courses.find((course) => course.code === courseToMove.code),
+    );
+    if (isDuplicate) {
+      toast.warn(
+        `You're already taking ${courseToMove.code} in ${destinationSemester.code.year}${destinationSemester.code.semester}`,
+      );
+      return;
+    }
+
+    setSemesters((semesters) =>
+      semesters.map((semester) => {
+        if (semester.id === destinationSemester.id) {
+          return { ...semester, courses: [...semester.courses, courseToMove] };
+        }
+
+        if (semester.id === originSemester.id) {
+          return {
+            ...semester,
+            courses: semester.courses.filter((course) => course.id !== courseToMove.id),
+          };
+        }
+        return semester;
+      }),
+    );
+
+    const oldSemesterId = originSemester.id.toString();
+    const newSemesterId = destinationSemester.id.toString();
+    const courseName = courseToMove.code;
+
+    addTask({ func: handleMoveCourse, args: { oldSemesterId, newSemesterId, courseName } });
+  };
 
   return (
     <DndContext
@@ -110,30 +331,19 @@ export default function Planner({
           // from course list -> semester
           if (
             originData.from === 'course-list' &&
-            destinationData.to === 'semester-tile' &&
-            onAddCourseToSemester
+            destinationData.to === 'semester-tile'
+            // onAddCourseToSemester
           ) {
-            onAddCourseToSemester(destinationData.semester, originData.course).then(
-              (notification) =>
-                // TODO: push message to toast notifications
-                console.log(notification.message),
-            );
+            handleAddCourseToSemester(destinationData.semester, originData.course);
           }
 
           // from semester -> another semester
-          if (
-            originData.from === 'semester-tile' &&
-            destinationData.to === 'semester-tile' &&
-            onMoveCourseFromSemesterToSemester
-          ) {
-            onMoveCourseFromSemesterToSemester(
+          if (originData.from === 'semester-tile' && destinationData.to === 'semester-tile') {
+            handleMoveCourseFromSemesterToSemester(
               originData.semester,
               destinationData.semester,
               originData.course,
-            ).then((notification) => {
-              // TODO: push message to toast notifications
-              console.log(notification.message);
-            });
+            );
           }
         }
       }}
@@ -163,10 +373,7 @@ export default function Planner({
               return (
                 <DroppableSemesterTile
                   onRemoveCourse={(semester, course) =>
-                    onRemoveCourseFromSemester(semester, course).then((notification) => {
-                      // TODO: push message to toast notification
-                      console.log(notification.message);
-                    })
+                    handleRemoveCourseFromSemester(semester, course)
                   }
                   key={semester.id.toString()}
                   dropId={`semester-${semester.id}`}
@@ -179,8 +386,8 @@ export default function Planner({
               );
             })}
             <div className="col-span-full flex justify-center items-center gap-8 h-10">
-              <button onClick={onRemoveYear}>- Remove Year</button>
-              <button onClick={onAddYear}>+ Add Year</button>
+              <button onClick={handleRemoveYear}>- Remove Year</button>
+              <button onClick={handleAddYear}>+ Add Year</button>
             </div>
           </div>
         </div>
